pyproject.toml
[project]
name = "driplink-lang-detector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
